+++ tests/Api/Groups/GroupTest.php
@@ -14,6 +14,7 @@ namespace Api\Groups;
 use Serhiy\Pushover\Api\Groups\Group;
 use PHPUnit\Framework\TestCase;
 use Serhiy\Pushover\Application;
+use Serhiy\Pushover\Client\Response\CreateGroupResponse;
 use Serhiy\Pushover\Client\Response\RetrieveGroupResponse;
 
 /**
@@ -65,4 +66,17 @@ class GroupTest extends TestCase
 
         $this->assertInstanceOf(RetrieveGroupResponse::class, $response);
     }
+    
+    /**
+     * @group Integration
+     */
+    public function testCreate()
+    {
+        $application = new Application("cccc3333CCCC3333dddd4444DDDD44"); // using dummy token
+        $group = new Group("eeee5555EEEE5555ffff6666FFFF66", $application);
+
+        $response = $group->create('unit test '. date('Y-m-d H:i:s'));
+
+        $this->assertInstanceOf(CreateGroupResponse::class, $response);
+    }    
 }