+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # Change Log
 
-
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
@@ -10,6 +9,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Fixed
 
 - Test setTtl() is duplicated #24
+
+### Added
+
+- Added support for creating a new group #26
 
 ## [1.4.0]